調整建議重補修名單，過濾一般、延修，並在畫面新增學生狀態。

+++ K12.Retake.Shinmin/DAO/GraduationPlanSubject.cs
@@ -30,7 +30,7 @@ namespace K12.Retake.Shinmin.DAO
         /// <summary>
         /// 學分數
         /// </summary>
-        public int Credit { get; set; }
+        public decimal Credit { get; set; }
 
         /// <summary>
         /// 必修或選修
